requirements.txt
Django
psycopg2-binary>=2.8
djangorestframework
djangorestframework-simplejwt
djoser
django-cors-headers
psycopg2
dj_database_url
